[ENH] `Structure` to `String` helper function (#61)
Waiting for #62 to be merged.

Solves #55 by adding a helper function `struct_to_aaseq` in `utils`. 

Location of helper function is open to feedback given I have no clear
idea where it should go. Just as a note, this uses `biopython` and will
be a very important base function in the development of this library.