Make PSeAAC more configurable (#126)
Makes `PSeAAC` more configurable by enabling customization of `lambda`
and `weight` parameters.

Main changes:
*  `lambda` and `weight` are now `__init__` parameters;
* updated a few docstrings here and there. Some were not consistent with
`numpydoc` (e.g., missing attributes, methods after examples, etc.);
* moved in-method imports at the top of the module.

A few things to mention. The output of `transform(...)` being of length
`20 + 30 = 50` was not a coincidence. In particular, the second part of
length 30 depends on the value of lambda. I updated the docstring
accordingly.

I noticed that `PSeAAC` has pretty bad test coverage, maybe due to #34
not being solved yet? I think we should open a separate issue and add a
comprehensive test suite for the algorithm so we don't forget about it.